requirements.txt
Flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
openai>=1.51.0
httpx==0.27.2
gunicorn==23.0.0
requests==2.32.3
PyPDF2==3.0.1
numpy==1.26.4
pandas==2.2.2
azure-storage-blob>=12.14.1
six==1.16.0
azure.search.documents==11.4.0